requirements.txt
flask
gunicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-cloud-secret-manager
gspread
requests
python-dotenv
